refactor: 에러 메시지 상수 파일명 변경
