+++ src/apis/shoppingCart/shoppingCart.ts
@@ -2,7 +2,7 @@ import { generateBasicToken } from '../../utils/auth';
 import { CartItem } from '../../type';
 
 import { fetchData } from '../';
-import { ERROR_MESSAGE } from '../fetchData/fetchData.constants';
+import { ERROR_MESSAGE } from '../fetchData/errorMessage';
 import { USER_ID, USER_PASSWORD, API_ROUTE } from './shoppingCart.constants';
 
 export async function fetchCartItems(): Promise<CartItem[]> {
